+++ lib/src/f_sp3d_material.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 import 'package:simple_3d/simple_3d.dart';
 
 /// (en)This class statically stores the default set of Sp3dMaterial that is often used.
+///
 /// (ja)Sp3dMaterialのよく使うデフォルトセットを静的に格納したクラスです。
 ///
 /// Author Masahide Mori
@@ -62,6 +63,60 @@ class FSp3dMaterial {
       true,
       0,
       const Color.fromARGB(0, 0, 0, 255));
+
+  static final Sp3dMaterial orange = Sp3dMaterial(
+      const Color.fromARGB(255, 255, 165, 0),
+      true,
+      1,
+      const Color.fromARGB(255, 255, 165, 0));
+
+  static final Sp3dMaterial orangeWire = Sp3dMaterial(
+      const Color.fromARGB(255, 255, 165, 0),
+      false,
+      1,
+      const Color.fromARGB(255, 255, 165, 0));
+
+  static final Sp3dMaterial orangeNonWire = Sp3dMaterial(
+      const Color.fromARGB(255, 255, 165, 0),
+      true,
+      0,
+      const Color.fromARGB(255, 255, 165, 0));
+
+  static final Sp3dMaterial pink = Sp3dMaterial(
+      const Color.fromARGB(255, 255, 192, 203),
+      true,
+      1,
+      const Color.fromARGB(255, 255, 192, 203));
+
+  static final Sp3dMaterial pinkWire = Sp3dMaterial(
+      const Color.fromARGB(255, 255, 192, 203),
+      false,
+      1,
+      const Color.fromARGB(255, 255, 192, 203));
+
+  static final Sp3dMaterial pinkNonWire = Sp3dMaterial(
+      const Color.fromARGB(255, 255, 192, 203),
+      true,
+      0,
+      const Color.fromARGB(255, 255, 192, 203));
+
+  static final Sp3dMaterial purple = Sp3dMaterial(
+      const Color.fromARGB(255, 160, 32, 240),
+      true,
+      1,
+      const Color.fromARGB(255, 160, 32, 240));
+
+  static final Sp3dMaterial purpleWire = Sp3dMaterial(
+      const Color.fromARGB(255, 160, 32, 240),
+      false,
+      1,
+      const Color.fromARGB(255, 160, 32, 240));
+
+  static final Sp3dMaterial purpleNonWire = Sp3dMaterial(
+      const Color.fromARGB(255, 160, 32, 240),
+      true,
+      0,
+      const Color.fromARGB(255, 160, 32, 240));
 
   static final Sp3dMaterial grey = Sp3dMaterial(
       const Color.fromARGB(255, 224, 224, 224),